refactor: merge topology submodules
