--- expertsystem/topology/__init__.py
@@ -1,9 +1,0 @@
-"""All modules related to topology building."""
-
-__all__ = [
-    "graph",
-    "topology_builder",
-]
-
-from . import graph
-from . import topology_builder
